Add user choice function
Added the function that asks for player input and returns the value
if it's valid. If it's not it reruns the function and returns.

When I use console.log(getUserChoice()) and the first value isn't valid
the console returns undefined but the actual value of the variable
changes. I don't quite understand why this works like this but I've got
the return I needed so I'm happy. Will look for an answer for this.