Add the play round function
For this function I needed to return the two values I was planning on
using later on the game function, and I quickly found out that
you can't. They say the best way would be to return the values inside an
array but I tried wrapping them in a string and it appears to work.
I also set the default values for the two score variables to 0 so the
increment operators work as intended.
I'm going to make the game function now and see if it all works.